Move trakt_list_util.py -> plex_trakt_sync/trakt_list_util.py
